requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.2.0
slither-analyzer>=0.10.0  # For Solidity parsing
tqdm>=4.65.0
pyyaml>=6.0.0
pytest>=7.3.0  # For tests